pyproject.toml
[project]
name = "srgan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "pillow>=12.0.0",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
]
